Repeated simulations with higher outcrossing rates based on new empirical data.
Changed the number of simulations from 500 to 2000, because populations converge much
quicker for higher outcrossing rates. I kept outcrossing rate of 5% to match the estimate
from Golenberg (1986) added an if statement to say that in this case run 2000
generations.

Rearrange file numbers and removed session info, because it wasn't clear which analysis
this was coming from. Also removed the log file and instead direct SLURM messages to a
folder (created in 002.sim_generator) with a message for each job.

 On branch main
 Changes to be committed:
	modified:   data_analysis/simulations/001.simulation_submission_script.sh
	modified:   data_analysis/simulations/002.sim_generator.R
	renamed:    data_analysis/simulations/005.summarise_simulations.R -> data_analysis/simulations/003.summarise_simulations.R
	deleted:    data_analysis/simulations/004.session_info.txt
	renamed:    data_analysis/simulations/006.simmiad_from_one_genotype.Rmd -> data_analysis/simulations/004.simmiad_from_one_genotype.Rmd

	modified:   README.md