Added simulations from structured population.
We want to simulate poulations that start with a 'realistic' structure, and see how
long that persists.

Added initial simulations to start from a structured population that runs simmiad with
`pop_structure='hardcoded'`, and supplies the vector of sampled DGGs from Ammiad along
the direction of a trnsect to be sampled. In each generation, the vector is first
rotated by a random number of positions. To generate a 2D structure, this copies
the vector of genotypes above and below to create bands of identical genotypes. That's
not ideal, but the other ideas I tried to simulate a 2D population were worse.

For a first try I used most-plausible parameter values (4% outcrossing, 1m dispersal,
30% seed bank, 4plants/m2) for 100 generations. This is done beginnning with the
structure in each of the sampling years, with 100 simulations per year.

 On branch main
 Changes to be committed:
	new file:   simulations/structured/001.sim_generator.R
	new file:   simulations/structured/002.simulation_submission_script.sh
	new file:   simulations/structured/003.summarise_simulations.R
	new file:   simulations/structured/004.submit_summarise_simulations.sh