Phenotypes work with dataset S1 and S2
Updated the phenotype analyses to work with the updated data files.
There was a bug that meant the variance explained by habitat came back
as zero for all traits. I think this was because I misspecified that
posterior_predict should draw phenotypes from levels in the vector of
genotypes. I confirmed that it was not because of differences in how
the genotypes themselves are labelled between datasets.

 On branch main
 Changes to be committed:
	modified:   data_analysis/phenotypes/001_linear_models.R
	modified:   data_analysis/phenotypes/002_submit_linear_models.sh
	modified:   data_analysis/phenotypes/003_heritability.R
	modified:   data_analysis/phenotypes/004_predict_phenotypes.R
	modified:   data_analysis/phenotypes/005_submit_predictions.sh
	modified:   data_analysis/phenotypes/006_plot_predictions.R
	modified:   data_processing/import_phenotype_data.R