Simulations track microhabitat differentiation
We changed and simplified the presentation of the simulations to start
with the observed structure at Ammiad in 1984 and track the change in
microhabitat differentiation through time.

In addition, in data_processing/ammiad_temporal_stability.R I specified
calls to dplyr::select with and explicit reference to package name,
because select is masked by the MASS package.

On branch main
 Changes to be committed:
	modified:   data_processing/ammiad_temporal_stability.R
	modified:   simulations/structured/001.sim_generator.R
	modified:   simulations/structured/002.simulation_submission_script.sh
	modified:   simulations/structured/003.summarise_simulations.R
	modified:   simulations/structured/005.plot_sims.R